training/requirements.txt
mlflow==2.12.2
scikit-learn==1.2.1
pyyaml==6.0.1
xgboost==1.7.6